locale.c: Revamp finding if locale is UTF-8
This changes how this functionality works for the LC_CTYPE locale.  On
systems that have nl_langinfo() one can get a definitive answer from
just that.  Otherwise (or if that doesn't return properly) one can use
mbtowc() to check if the UTF-8 byte sequence for the Unicode REPLACEMENT
CHARACTER actually is considered to be that code point.  This is also
definitive.  If the maximum byte string length for a character is too
short to handle all Unicode UTF-8, we know without further checking that
this isn't a UTF-8 locale, so can avoid the mbtowc check.