Use dfa to speed up translating UTF-8 into code point
This dfa is available from the internet has the reputation of being the
fastest general translator.  It does not allow Perl extended UTF-8.
This commit changes to use it at the beginning of our translator,
modifying it slightly to accept surrogates.  If necessary, it drops down
into our translator to handle errors and warnings and Perl extended.